Refactor and cleanup zwave_js fixtures to use
