+++ tests/components/hassio/common.py
@@ -23,19 +23,25 @@ def mock_discovery_info() -> Any:
     return DEFAULT
 
 
-def mock_get_addon_discovery_info(discovery_info: Any) -> Generator[AsyncMock]:
+def mock_get_addon_discovery_info(
+    discovery_info: dict[str, Any], discovery_info_side_effect: Any | None
+) -> Generator[AsyncMock]:
     """Mock get add-on discovery info."""
     with patch(
         "homeassistant.components.hassio.addon_manager.async_get_addon_discovery_info",
+        side_effect=discovery_info_side_effect,
         return_value=discovery_info,
     ) as get_addon_discovery_info:
         yield get_addon_discovery_info
 
 
-def mock_addon_store_info() -> Generator[AsyncMock]:
+def mock_addon_store_info(
+    addon_store_info_side_effect: Any | None,
+) -> Generator[AsyncMock]:
     """Mock Supervisor add-on store info."""
     with patch(
-        "homeassistant.components.hassio.addon_manager.async_get_addon_store_info"
+        "homeassistant.components.hassio.addon_manager.async_get_addon_store_info",
+        side_effect=addon_store_info_side_effect,
     ) as addon_store_info:
         addon_store_info.return_value = {
             "available": False,
@@ -46,10 +52,11 @@ def mock_addon_store_info() -> Generator[AsyncMock]:
         yield addon_store_info
 
 
-def mock_addon_info() -> Generator[AsyncMock]:
+def mock_addon_info(addon_info_side_effect: Any | None) -> Generator[AsyncMock]:
     """Mock Supervisor add-on info."""
     with patch(
         "homeassistant.components.hassio.addon_manager.async_get_addon_info",
+        side_effect=addon_info_side_effect,
     ) as addon_info:
         addon_info.return_value = {
             "available": False,
@@ -96,18 +103,18 @@ def mock_addon_running(addon_store_info: AsyncMock, addon_info: AsyncMock) -> As
         "version": "1.0.0",
     }
     addon_info.return_value["available"] = True
-    addon_info.return_value["hostname"] = "core-mosquitto"
+    addon_info.return_value["hostname"] = "core-test-addon"
     addon_info.return_value["state"] = "started"
     addon_info.return_value["version"] = "1.0.0"
     return addon_info
 
 
-def mock_install_addon(
+def mock_install_addon_side_effect(
     addon_store_info: AsyncMock, addon_info: AsyncMock
-) -> Generator[AsyncMock]:
-    """Mock install add-on."""
+) -> Any | None:
+    """Return the install add-on side effect."""
 
-    async def install_addon_side_effect(hass: HomeAssistant, slug: str) -> None:
+    async def install_addon(hass: HomeAssistant, slug):
         """Mock install add-on."""
         addon_store_info.return_value = {
             "available": True,
@@ -119,17 +126,43 @@ def mock_install_addon(
         addon_info.return_value["state"] = "stopped"
         addon_info.return_value["version"] = "1.0.0"
 
+    return install_addon
+
+
+def mock_install_addon(install_addon_side_effect: Any | None) -> Generator[AsyncMock]:
+    """Mock install add-on."""
+
     with patch(
-        "homeassistant.components.hassio.addon_manager.async_install_addon"
+        "homeassistant.components.hassio.addon_manager.async_install_addon",
+        side_effect=install_addon_side_effect,
     ) as install_addon:
-        install_addon.side_effect = install_addon_side_effect
         yield install_addon
 
 
-def mock_start_addon() -> Generator[AsyncMock]:
+def mock_start_addon_side_effect(
+    addon_store_info: AsyncMock, addon_info: AsyncMock
+) -> Any | None:
+    """Return the start add-on options side effect."""
+
+    async def start_addon(hass: HomeAssistant, slug):
+        """Mock start add-on."""
+        addon_store_info.return_value = {
+            "available": True,
+            "installed": "1.0.0",
+            "state": "started",
+            "version": "1.0.0",
+        }
+        addon_info.return_value["available"] = True
+        addon_info.return_value["state"] = "started"
+
+    return start_addon
+
+
+def mock_start_addon(start_addon_side_effect: Any | None) -> Generator[AsyncMock]:
     """Mock start add-on."""
     with patch(
-        "homeassistant.components.hassio.addon_manager.async_start_addon"
+        "homeassistant.components.hassio.addon_manager.async_start_addon",
+        side_effect=start_addon_side_effect,
     ) as start_addon:
         yield start_addon
 
@@ -142,10 +175,11 @@ def mock_stop_addon() -> Generator[AsyncMock]:
         yield stop_addon
 
 
-def mock_restart_addon() -> Generator[AsyncMock]:
+def mock_restart_addon(restart_addon_side_effect: Any | None) -> Generator[AsyncMock]:
     """Mock restart add-on."""
     with patch(
-        "homeassistant.components.hassio.addon_manager.async_restart_addon"
+        "homeassistant.components.hassio.addon_manager.async_restart_addon",
+        side_effect=restart_addon_side_effect,
     ) as restart_addon:
         yield restart_addon
 
@@ -158,10 +192,28 @@ def mock_uninstall_addon() -> Generator[AsyncMock]:
         yield uninstall_addon
 
 
-def mock_set_addon_options() -> Generator[AsyncMock]:
+def mock_addon_options(addon_info: AsyncMock) -> dict[str, Any]:
+    """Mock add-on options."""
+    return addon_info.return_value["options"]
+
+
+def mock_set_addon_options_side_effect(addon_options: dict[str, Any]) -> Any | None:
+    """Return the set add-on options side effect."""
+
+    async def set_addon_options(hass: HomeAssistant, slug: str, options: dict) -> None:
+        """Mock set add-on options."""
+        addon_options.update(options["options"])
+
+    return set_addon_options
+
+
+def mock_set_addon_options(
+    set_addon_options_side_effect: Any | None,
+) -> Generator[AsyncMock]:
     """Mock set add-on options."""
     with patch(
-        "homeassistant.components.hassio.addon_manager.async_set_addon_options"
+        "homeassistant.components.hassio.addon_manager.async_set_addon_options",
+        side_effect=set_addon_options_side_effect,
     ) as set_options:
         yield set_options
