+++ tests/components/zwave_js/test_config_flow.py
@@ -5,7 +5,7 @@ from collections.abc import Generator
 from copy import copy
 from ipaddress import ip_address
 from typing import Any
-from unittest.mock import DEFAULT, AsyncMock, MagicMock, call, patch
+from unittest.mock import AsyncMock, MagicMock, call, patch
 
 import aiohttp
 import pytest
@@ -75,31 +75,6 @@ def mock_supervisor_fixture() -> Generator[None]:
         "homeassistant.components.zwave_js.config_flow.is_hassio", return_value=True
     ):
         yield
-
-
-@pytest.fixture(name="discovery_info")
-def discovery_info_fixture() -> dict[str, Any]:
-    """Return the discovery info from the supervisor."""
-    return DEFAULT
-
-
-@pytest.fixture(name="discovery_info_side_effect")
-def discovery_info_side_effect_fixture() -> Any | None:
-    """Return the discovery info from the supervisor."""
-    return None
-
-
-@pytest.fixture(name="get_addon_discovery_info")
-def mock_get_addon_discovery_info(
-    discovery_info: dict[str, Any], discovery_info_side_effect: Any | None
-) -> Generator[AsyncMock]:
-    """Mock get add-on discovery info."""
-    with patch(
-        "homeassistant.components.hassio.addon_manager.async_get_addon_discovery_info",
-        side_effect=discovery_info_side_effect,
-        return_value=discovery_info,
-    ) as get_addon_discovery_info:
-        yield get_addon_discovery_info
 
 
 @pytest.fixture(name="server_version_side_effect")
